[ADDED] Event Page

+++ src/pages/dashboard/Events.jsx
@@ -3,9 +3,12 @@ import {empty} from '@utils/helpers';
 import React, {useState} from 'react';
 import {BiEditAlt} from 'react-icons/bi';
 import {getEvents} from '@services/Events';
+import Permissions from '@utils/Permissions';
 import {AiOutlinePlus} from 'react-icons/ai';
 import {HiOutlineTrash} from 'react-icons/hi';
+import { useAuth } from '@context/AuthContext';
 import useEvents from '@hooks/Dashboard/useEvents';
+import useCatalogs from '@hooks/Dashboard/useCatalogs';
 import CustomHeader from '@components/UI/CustomHeader';
 import Breadcrumb from '@components/Dashboard/Breadcrumb';
 import CustomModal from '@components/UI/Modal/CustomModal';
@@ -14,9 +17,11 @@ import DeleteDialog from '@components/Dashboard/DeleteDialog';
 import EventForm from '@components/Dashboard/Event/EventForm';
 
 const Events = () => {
+  const {permissions} = useAuth();
+  const {faculties} = useCatalogs();
   const [eventIdToDelete, setEventIdToDelete] = useState(null);
   const {onEdit, onCreate, onUpdate, onDelete, onCloseForm, onToggleForm, isOpenForm, currentData} = useEvents();
-  const customActions = getCustomActions(onEdit, setEventIdToDelete);
+  const customActions = getCustomActions(onEdit, setEventIdToDelete, permissions.includes(Permissions.MANAGE_EVENTS));
 
   return (
     <div>
@@ -26,14 +31,16 @@ const Events = () => {
 
       <div className="flex justify-between items-center mb-4 mt-1">
         <h1 className="text-primary text-3xl font-bold">Eventos</h1>
-        <button
-          type="button"
-          onClick={onToggleForm}
-          className="btn flex justify-center items-center bg-primary text-white px-4 py-2.5 rounded-lg hover:bg-secondary"
-        >
-          <AiOutlinePlus className="mr-2" />
-          <span>Agregar evento</span>
-        </button>
+        {permissions.includes(Permissions.MANAGE_EVENTS) && (
+          <button
+            type="button"
+            onClick={onToggleForm}
+            className="btn flex justify-center items-center bg-primary text-white px-4 py-2.5 rounded-lg hover:bg-secondary"
+          >
+            <AiOutlinePlus className="mr-2" />
+            <span>Agregar evento</span>
+          </button>
+        )}
       </div>
 
       <CustomTable
@@ -51,38 +58,46 @@ const Events = () => {
         onDelete={() => onDelete.mutate(eventIdToDelete)}
       />
 
-      <CustomModal
-        isOpen={isOpenForm}
-        onToggleModal={onCloseForm}
-        className="p-0 w-full sm:max-w-3xl"
-      >
-        <EventForm
-          initialData={currentData}
-          onCreate={onCreate}
-          onUpdate={onUpdate}
-          onClose={onCloseForm}
-        />
-      </CustomModal>
+      {permissions.includes(Permissions.MANAGE_EVENTS) && (
+        <CustomModal
+          isOpen={isOpenForm}
+          onToggleModal={onCloseForm}
+          className="p-0 w-full sm:max-w-3xl"
+        >
+          <EventForm
+            initialData={currentData}
+            onCreate={onCreate}
+            onUpdate={onUpdate}
+            onClose={onCloseForm}
+            faculties={faculties}
+          />
+        </CustomModal>
+      )}
     </div>
   );
 };
 
-const getCustomActions = (onEdit, onDelete) => [
-  {
-    id: 1,
-    label: '',
-    tooltip: 'Editar',
-    Icon: BiEditAlt,
-    onClick: onEdit,
-  },
-  {
-    id: 2,
-    label: '',
-    tooltip: 'Borrar',
-    Icon: HiOutlineTrash,
-    onClick: (data) => onDelete(data.id),
-  },
-];
+const getCustomActions = (onEdit, onDelete, userHasPermissionsToManage) => {
+
+  if (!userHasPermissionsToManage) return [];
+
+  return [
+    {
+      id: 1,
+      label: '',
+      tooltip: 'Editar',
+      Icon: BiEditAlt,
+      onClick: onEdit,
+    },
+    {
+      id: 2,
+      label: '',
+      tooltip: 'Borrar',
+      Icon: HiOutlineTrash,
+      onClick: (data) => onDelete(data.id),
+    },
+  ]
+};
 
 const columns = [
   {
@@ -91,15 +106,20 @@ const columns = [
   },
   {
     title: 'Fecha de inicio',
-    field: 'startDate',
+    field: 'formatStartDate',
     stackedColumn: true,
     className: 'hidden lg:table-cell',
   },
   {
     title: 'Fecha de fin',
-    field: 'endDate',
+    field: 'formatEndDate',
     stackedColumn: true,
     className: 'hidden lg:table-cell',
+  },
+  {
+    title: 'Facultades',
+    field: 'permissions',
+    render: (rowData) => <Faculties data={rowData.faculties} />,
   },
   {
     title: 'Capacidad',
@@ -107,6 +127,36 @@ const columns = [
     stackedColumn: true,
     className: 'hidden lg:table-cell',
   },
+  {
+    title: 'Estado',
+    field: 'enabled',
+    render: (rowData) => <BadgeStatus status={rowData.enabled} />,
+  },
 ];
+
+const Faculties = ({data}) => {
+  return (
+    <ul className="list-disc list-inside">
+      {data.map((faculty) => (
+        <li key={faculty.value} className="text-xs">{faculty.name}</li>
+      ))}
+    </ul>
+  )
+}
+
+const BadgeStatus = ({status}) => {
+  const customClassName =
+    status
+      ? 'bg-green-100 text-green-600'
+      : 'bg-red-100 text-red-600';
+
+  return (
+    <div
+      className={`flex justify-center item-center bg-green-100 px-3 py-2 rounded-lg ${customClassName}`}
+    >
+      <p className="font-bold">{status ? 'Activo' : 'Inactivo'}</p>
+    </div>
+  );
+};
 
 export default Events;
